straighten out socket file mess, part 1
This was the hard part. Under abnormal termination (kill -9,
seg faults, whatever) socket files do not get cleaned up properly.
As long as this is the IPC mechanism, there is very little that
can be done about this. So, with this change we are leaving them
to linger. Fixing this we can do one of two things. First,
we can look for EADDRINUSE and just delete the file and retry.
However, that prevents someone from running two binaries at the
same time (maybe to run two servers on different ports). The
second option is to make the files unique, so this is the path
chosen. Here we append a system timestamp to the file name.

Part two of this eventually will need to be locating the TEMP
directory and placing the files there.